requirements.txt
python-multipart
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
pyarrow
lightgbm
shap
requests
streamlit
passlib[bcrypt]
python-jose
python-dotenv
pika
plotly
# transformers   # optional fallback
# torch          # optional: huge, remove for faster builds if not used
